refactor: Merge the two SearchProviders
It doesn't seem useful to have a wrapper
